server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Miloš Marinković <milos.marinkovic@biosens.rs>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.2"
flask-cors = "^4.0.0"
config = "^0.5.1"
python-dotenv = "^1.0.1"
marshmallow = "^3.21.1"
rdflib = "^7.0.0"
pypdf2 = "^3.0.1"
openai = "0.28"
pandas = "^2.2.1"
unidecode = "^1.3.8"
annoy = "^1.17.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
